[close #358] Fail on Rake detect for Rails Apps
Rails applications should be reasonably expected to run `rake` when a Rakefile is present.

Right now this is the behavior we support https://github.com/heroku/heroku-buildpack-ruby/issues/358#issuecomment-90250883. While it might be reasonable that a Ruby application might not expect us to run Rake, a Rails app with the asset pipeline, while they might not be using the asset pipeline, should be able to run the `rake` command when a Rakefile is present.