Warn if ruby/rails/bundle binstub shebangs are bad
Binstubs, which RubyGems and Bundler generate to prepare the environment
before dispatching calls to original executables, can sometimes include
shebangs that worked on developer machines, but won't work on Heroku.
Often, this is caused by ruby binaries being named other than `ruby` (or
`ruby.exe` on Windows).

While it can happen for any binstub, we're mostly concerned about
`rake`, `rails`, and `bundler` stubs. Check each of these to see if they
are `!#/usr/bin/env ruby` or `!#/usr/bin/env ruby.exe` and issue a
warning if this is not the case.