Fix double newline printing
Introduced in #724 the `puts` method now uses `print` but appends a newline if it’s not already present. Unfortunately the string in the `end_with?` call is using single quotes instead of double quotes, so it is checking for a string that literally ends with a slash and an n rather than the newline character.

Switching to double quotes fixes the issue.