Do not expect `rails runner` to be error free
When people have exceptions in their Rails apps that only show on boot, previously the app would deploy and then they would have to debug in production. This call to `rails runner` preemptively calls code that might not be called until after build and so people are getting exceptions here. It’s confusing, so this warning and check are now optional. We will attempt to run them and give an appropriate warning when possible, otherwise we will not warn.