Fix Rails Runner timeout
IO.popen cannot be timed out reliably: https://stackoverflow.com/questions/17237743/timeout-within-a-popen-works-but-popen-inside-a-timeout-doesnt

We could instead use exec (i.e. backticks) and we can time it out. However, there’s a problem because this code:

```
require 'timeout'

begin
  Timeout.timeout(1) do
    `ruby -e 'pid = Process.spawn("sleep 1000"); Process.wait(pid)'`
  end
rescue Timeout::Error
end
```

This code leaves a process lying around even though the parent process exits. The Heroku build system interprets a running process as a thing that we need to wait for and so the build never ends.

To meet our needs we need:

- To be able to reliably time out this command.
- To provide debugging output for when the command fails or times out.
- Not leave hanging processes around when the buildpack ends.

One way that we can time out and not leave processes hanging is via `Process.spawn`

```
require 'timeout'

begin
  Timeout.timeout(1) do
    @pid = Process.spawn(%Q{`ruby -e 'pid = Process.spawn("sleep 1000"); Process.wait(pid)'`})
    Process.wait(@pid)
  end
rescue Timeout::Error
  Process.kill("SIGTERM", @pid)
end
```

When the timeout fires, the child process is killed.

Now that we can stop a process, we need to be able to also log the output for debugging.

The solution that I came up with is to tell the command to write to a file while it is running. If the command is killed via a timeout then the contents of the file are preserved and can be used for debugging. 
The updated timeout test fails on master, and passes on this branch.