Close #783 Warn on `action_dispatch.x_sendfile_header`
When serving files if the `action_dispatch.x_sendfile_header` is set then it is essentially saying “i don’t need to serve this asset, let nginx or apache do it for me”. The problem is…on heroku most people aren’t using apache or nginx so this causes a silent failure that is hard to debug.

This warning will hopefully help people identify their mistake earlier.