Fix a flappy test based on the bundler wrapper
Usually this would take a LONG time as the test suite when run serially takes about an hour. The other thing that is tough about this, is it only fails on CI and not locally.

On a hunch I guessed the failing behavior came only from `spec/helpers` which is all unit tested and therefore pretty fast. This allowed me to actually find the failing order.

Via `$ heroku ci:debug`

First I ran:

```
$ bundle exec rspec spec/helpers --bisect --seed 1
```

The `--seed` argument is used because otherwise the order is randomized on every run. It doesn't matter what value is used as long as it's consistent.

This produce a "minimal failing example" of running:

```
$ rspec './spec/helpers/fetcher_spec.rb[1:1]' './spec/helpers/rails_runner_spec.rb[1:3,1:5]' './spec/helpers/rake_runner_spec.rb[1:1,1:2]' --seed 1
```

Sure enough this command fails 100% of the time.

I knew the culprit had to be mutating something global. Previously I've been focused on files on disk but considering every example is now run inside of it's own temp directory then it is likely something else. The other culprit could be an environment variable.

This lead me to look into the bundler wrapper which sure enoug mutates the env var `BUNDLE_GEMFILE`.