[changelog skip] Fix logic for Outdated check
Previously when calling this logic on a Ruby version that is truly the last in line, then we would end up with an empty array that we then call `first` on and then expect it to be a string. That string we then call `sub` on. This fails when nil is returned from the array.

Instead of doing this logic inside of the join method, we can put it directly in the suggest_ruby_eol_version method which runs after a guard statement so it will never run unless the array provided a value.

We can test this in a way that does not depend on external existence of the files by stubbing out the fetcher to always return false when it checks for another ruby version.