Fix absolute path bug in .profile.d/ruby.sh $PATH
In #888 a relative path was accidentally introduced into the `.profile.d/ruby.sh` script:

```
export PATH="bin:$HOME/vendor/bundle/bin:$HOME/vendor/bundle/ruby/2.6.0/bin:$PATH"
```

That first path `bin` is relative and should be `$HOME/bin`.

Without it, then changing into a different directory means that the system version of ruby will be used

```
~ $ which ruby
bin/ruby
~ $ mkdir foo; cd foo
~/foo $ which ruby
/usr/bin/ruby
```

This is a regression  was not caught by tests previously because the version of ruby that the test `cd_ruby` uses is 2.5.1 and that happens to match system ruby. To fix for catching future regressions the test is now more explicit.

Unfortunately there's a bug in hatchet preventing that specific test https://github.com/heroku/hatchet/issues/71 so instead I'm at least asserting that it's an absolute path.