Handle binary binstubs (#1021)
* Revert "Merge pull request #1020 from heroku/schneems/revert-bad-shebang-check"

This reverts commit da07466ec623aa4312c9f07df1777f93b780b156, reversing
changes made to aeed0ce5e0977157060c010b502b728480ed2047.

* Handle binary binstubs

Logic was added to check the "shebang" lines of binstubs. This was then reverted due to a bug #1019. This commit fixes this error by checking if a shebang line can be considered UTF-8 for purposes of comparison before we compare it. 

In the process I also refactored the code so that the logic around working with the binstub file is in BinstubWrapper. 

I added a test for the `BinstubWrapper#binary?` method as well as a higher level test that runs the whole check against the directory that contains the `which ruby` binary.

* Document v216 rollback in git

This deploy was rolled back due to #1005