[changelog skip] Promote tag to release and vendor CNB (#1044)
* [changelog skip] Promote tag to release and vendor CNB

When a tag is created, this action will build a vendored version of the Ruby buildpack for CNB, create a release from the tag, and then attach the tarball of the buildpack to the release.

To cut a new release run:

```
$ git tag v219 && git push --tags
```

* Move tarball logic to class for testing

Logic in the rake file is not tested. I want to start porting code over to custom classes that can be tested. Then we can use those classes with a minimal interface in the rakefile. It won't guarantee there are no regressions, but it will give much more confidence when modifying the logic.

In addition to moving this logic over to a custom class. The logic is changed so that it's tarballing contents from disk instead of from git remote so it can be fired from GitHub actions.

* Add `rake buildpack:release` command

This command will run checks to ensure that the local copy is the same as the copy on GitHub. It then creates and pushes a tag to GitHub, then uses the heroku buildpack CLI to deploy the change.

Currently the process for deploying a buildpack involves manually making a tag, then calling the cli. This allows us to turn that into a one step process that we can add guards to (for instance checking changelog contents).

The majority of the logic for this rake task is in a custom class that is tested.