Update default Ruby version to 3.0.3 (#1270)
* Update default Ruby version to 3.0.3

## CircleCI Ruby version

I'm unable to get RVM working with 3.0.3 for some reason on circleci. We need an exact ruby version, not just any 3.0.x version so their images won't work. I'm switching over to ruby-install and chruby for total control. I first tried installing via apt-get and for some reason it couldn't find them even after running an apt-get update.

This is gnarly, but it works and gives me a lot of control. It's frustrating to have to go down this route, but I feel good about the final solution

## CircleCI docker setup

Cutlass interfaces with docker through the API and needs a different setup otherwise it fails with error:

```
       stderr: ERROR: downloading buildpack from docker://cutlass_local_buildpack_21223a55e82b762fde3c: unsupported protocol docker in URI docker://cutlass_local_buildpack_21223a55e82b762fde3c
     # /home/circleci/.gem/ruby/3.0.3/gems/cutlass-0.3.0/lib/cutlass/pack_build.rb:96:in `call_pack'
     # /home/circleci/.gem/ruby/3.0.3/gems/cutlass-0.3.0/lib/cutlass/pack_build.rb:73:in `call'
     # /home/circleci/.gem/ruby/3.0.3/gems/cutlass-0.3.0/lib/cutlass/app.rb:118:in `pack_build'
     # ./spec/cnb/basic_local_pack_spec.rb:88:in `block (3 levels) in <top (required)>'
```

## Ruby 3 syntax

I hit some issues with kwargs and super. Using `**` instead of `*`

## CNB runner

The home-brew cnb runner was failing and I don't know why. Switching to cutlass fixed the issues. It wasn't worth it to investigate more.

* Update .circleci/config.yml

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

* Update spec/spec_helper.rb

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

* Update .circleci/config.yml

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>

* Remove incorrect caching code

https://github.com/heroku/heroku-buildpack-ruby/pull/1270#discussion_r809501218

* Fix tests dependent on heroku-18 stack

Co-authored-by: Ed Morley <501702+edmorley@users.noreply.github.com>